First rough tensorberg code base.
Need to shrink model to do toy training.

+++ data_loader/data_utils.py
@@ -40,7 +40,7 @@ def load_CIFAR_batch(filename):
         datadict = load_pickle(f)
         X = datadict['data']
         Y = datadict['labels']
-        X = X.reshape(10000, 3, 32, 32).transpose(0,2,3,1).astype("float")
+        X = X.reshape(10000, 3, 32, 32).transpose(0,2,3,1).astype("float32")
         Y = np.array(Y)
         return X, Y
